+++ rust/main/lander/src/adapter/chains/ethereum/nonce/state/validate/tests.rs
@@ -3,50 +3,12 @@ use ethers_core::types::Address;
 use hyperlane_core::U256;
 
 use crate::tests::test_utils::tmp_dbs;
-use crate::transaction::{
-    DropReason, Transaction, TransactionStatus, TransactionUuid, VmSpecificTxData,
-};
+use crate::transaction::{DropReason, TransactionStatus, TransactionUuid};
 use crate::TransactionDropReason;
 
-use super::super::super::super::EthereumTxPrecursor;
+use super::super::tests::make_tx;
 use super::NonceAction;
 use super::NonceManagerState;
-
-#[allow(deprecated)]
-fn make_tx(
-    uuid: TransactionUuid,
-    status: TransactionStatus,
-    nonce: Option<U256>,
-    address: Option<Address>,
-) -> Transaction {
-    use ethers_core::abi::Function;
-    let mut precursor = EthereumTxPrecursor {
-        tx: Default::default(),
-        function: Function {
-            name: "".to_string(),
-            inputs: vec![],
-            outputs: vec![],
-            constant: None,
-            state_mutability: Default::default(),
-        },
-    };
-    if let Some(n) = nonce {
-        precursor.tx.set_nonce(n);
-    }
-    if let Some(addr) = address {
-        precursor.tx.set_from(addr);
-    }
-    Transaction {
-        uuid,
-        tx_hashes: vec![],
-        vm_specific_data: VmSpecificTxData::Evm(precursor),
-        payload_details: vec![],
-        status,
-        submission_attempts: 0,
-        creation_timestamp: Default::default(),
-        last_submission_attempt: None,
-    }
-}
 
 #[tokio::test]
 async fn test_validate_assigned_nonce_none_nonce() {
